helpers: move dkim types
